Scope pytest fixtures
Our tests are becoming significantly slow because we are not setting a proper scope for our pytest fixtures. With the new change, the time for nox session `test-3.8(airflow='2.5')` has  been reduced from ~6m down to ~2m i.e. a 3x speed improvement.
